+++ src/Products/PloneMeeting/vocabularies.py
@@ -1205,22 +1205,28 @@ class FacetedAnnexesVocabulary(object):
         """ """
         tool = api.portal.get_tool('portal_plonemeeting')
         cfg = tool.getMeetingConfig(context)
-        annexes_config = cfg.annexes_types.item_annexes
+        annexes_configs = [annex_config for annex_config in
+                           cfg.annexes_types.objectValues()]
         config = OrderedDict([
             ('to_be_printed_activated', ("to_print", "not_to_print")),
             ('confidentiality_activated', ("confidential", "not_confidential")),
             ('publishable_activated', ("publishable", "not_publishable")),
             ('signed_activated', ("to_sign", "not_to_sign", "signed"))])
         res = []
+        values_enabled = []
         for k, values in config.items():
-            if getattr(annexes_config, k, False) is True:
-                for value in values:
-                    res.append(SimpleTerm(
-                        value,
-                        value,
-                        translate('annex_term_{0}'.format(value),
-                                  domain='PloneMeeting',
-                                  context=context.REQUEST)))
+            for annexes_config in annexes_configs:
+                if getattr(annexes_config, k, False) is True:
+                    for value in values:
+                        if value not in values_enabled:
+                            values_enabled.append(value)
+        for value in values_enabled:
+            res.append(SimpleTerm(
+                value,
+                value,
+                translate('annex_term_{0}'.format(value),
+                          domain='PloneMeeting',
+                          context=context.REQUEST)))
         return SimpleVocabulary(res)